better example in README.md
The "full message" example was using a fixed date/timestamp from 2010. While this works, it does not show up in the default Graylog stream views as it is from the distant past, which made me think that this module was broken or not sending the message. Fix this by using the current timestamp.